+++ src/f64x4_.rs
@@ -223,6 +223,7 @@ impl_simd_float! {
     N = 4,
     Simd = f64x4,
     UnsignedT = u64,
+    UnsignedSimd = u64x4,
   }
   old_powf_simd_fn_name = pow_f64x4,