+++ src/u16x16_.rs
@@ -163,6 +163,7 @@ impl_simd_uint! {
     T = u16,
     N = 16,
     Simd = u16x16,
+    SignedSimd = i16x16,
     T_BITS = 16,
     T_BITS_MUL_2 = 32,
     [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15],