+++ src/u32x16_.rs
@@ -166,6 +166,7 @@ impl_simd_uint! {
     T = u32,
     N = 16,
     Simd = u32x16,
+    SignedSimd = i32x16,
     T_BITS = 32,
     T_BITS_MUL_2 = 64,
     [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15],