+++ src/u64x2_.rs
@@ -231,6 +231,7 @@ impl_simd_uint! {
     T = u64,
     N = 2,
     Simd = u64x2,
+    SignedSimd = i64x2,
     T_BITS = 64,
     T_BITS_MUL_2 = 128,
     [0, 1],