+++ src/u8x16_.rs
@@ -337,6 +337,7 @@ impl_simd_uint! {
     T = u8,
     N = 16,
     Simd = u8x16,
+    SignedSimd = i8x16,
     T_BITS = 8,
     T_BITS_MUL_2 = 16,
     [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15],